IBX-538: Renamed EzPublishIO Bundle & Extension to IbexaIO
